#18 #16 Converted to use posts as routes, show individual post, update post fields
