+++ models/Post.js
@@ -1,22 +1,38 @@
 const mongoose = require('mongoose')
 
 const PostSchema = new mongoose.Schema({
-  post: {
+  title: {
     type: String,
     required: true,
   },
-  caption: {
+  image: {
     type: String,
     required: true
   },
-  image: {
+//   cloudinaryId: {
+//     type: String,
+//     require: true,
+//   },
+  caption: {
     type: String,
     required: true
   },
   link: {
     type: String,
     required: false
-  }
+  },
+//   likes: {
+//     type: Number,
+//     required: true,
+//   },
+  user: {
+    type: mongoose.Schema.Types.ObjectId,
+    ref: "User",
+  },
+  createdAt: {
+    type: Date,
+    default: Date.now,
+  },
 })
 
 module.exports = mongoose.model('Post', PostSchema)
