+++ routes/main.js
@@ -2,9 +2,12 @@ const express = require('express')
 const router = express.Router()
 const authController = require('../controllers/auth') 
 const homeController = require('../controllers/home')
+const postsController = require("../controllers/posts");
 const { ensureAuth, ensureGuest } = require('../middleware/auth')
 
 router.get('/', homeController.getIndex)
+router.get("/profile", ensureAuth, postsController.getProfile);
+//router.get("/feed", ensureAuth, postsController.getFeed);
 router.get('/login', authController.getLogin)
 router.post('/login', authController.postLogin)
 router.get('/logout', authController.logout)
