+++ routes/profile.js
@@ -1,12 +1,12 @@
-const express = require('express')
-const router = express.Router()
-const profileController = require('../controllers/profile') 
-const { ensureAuth, ensureGuest } = require('../middleware/auth')
+// const express = require('express')
+// const router = express.Router()
+// const profileController = require('../controllers/profile') 
+// const { ensureAuth, ensureGuest } = require('../middleware/auth')
 
-router.get('/', ensureAuth, profileController.getPost)
+// router.get('/', ensureAuth, profileController.getPost)
 
-router.post('/createPost', profileController.createPost)
+// router.post('/createPost', profileController.createPost)
 
-//router.delete('/deletePost', profileController.deletePost)
+// //router.delete('/deletePost', profileController.deletePost)
 
-module.exports = router
+// module.exports = router
