+++ views/post.ejs
@@ -1,0 +1,35 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <%- include('../views/partials/head'); %>
+</head>
+<body>
+    <header>
+        <%- include('../views/partials/header'); %>
+    </header>
+
+    <h2>
+        <%= post.title %>
+    </h2>
+    <p class="image"><img class="img-fluid" src="<%= post.image%>" /></p>
+
+<div>
+<%if(post.user==user.id){ %>
+<form action="/post/deletePost/<%= post.id %>?_method=DELETE" method="POST" class="col-3">
+<button class="btn btn-primary fa fa-trash" type="submit"></button>
+</form>
+<%}%>
+</div>
+    <div class="col-3 mt-5">
+        <p>
+            <%= post.caption %>
+        </p>
+    </div>
+    <div class="col-6 mt-5">
+        <a class="btn btn-primary" href="/profile">Return to Profile</a>
+        <a class="btn btn-primary" href="/feed">Return to Feed</a>
+    </div>
+
+    <script src="js/main.js"></script>
+</body>
+</html>
